Desktop: break up the code
